aa/backend/user_service/requirements.txt
Flask
Flask-Cors
requests
python-dotenv
supabase